# Add primitive fast path for ArrayConcatFunction